Move sample metadata fields to parallel hdf5 datasets
The previous implementation would overwrite the metadata attributes each time
a new raster was parsed; this version allows multiple versions to exist in
parallel. The metadata itself is tied to each sample using an index that
corresponds to the position of the metadata string in the separate dataset.
This implementation also stores the entire raster YAML metadata dict as a
single string that may be eval'd and reinstantiated as needed at runtime.