Refactor image reader & vector rasterizer utilities
The current version of these functions is now more generic than before. The rasterization
utility function (vector_to_raster) is now located in the 'utils' package and supports the
burning of vectors into separate layers as well as in the same layer (which is the original
behavior). The new multi-layer behavior is used in the updated 'image_reader_as_array' utility
function to (optionally) append new layers to raw imagery.

The refactoring also allowed the cleanup of the 'assert_band_number' utility function, and
simplification of the code in both the inference script (inference.py') and dataset preparation
script ('image_to_samples.py').